chore(utils/storage): rename localStorage with storage
